refactor(Deck.java) : 덱 자료구조 Deque로 변경 및 리팩토링
- foreach 부분 메소드 추출
- 카드 2개 뽑을 때 Stream.generate()로 수정
