+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,32 +1,32 @@
 package blackjack.domain.card;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Stack;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class Deck {
-    private final Stack<Card> cards;
+    private final Deque<Card> cards;
 
     public Deck() {
         cards = init();
     }
 
-    private Stack<Card> init() {
-        Stack<Card> cards = new Stack<>();
+    private Deque<Card> init() {
+        List<Card> cards = new ArrayList<>();
         Arrays.stream(Suit.values())
-            .forEach(suit -> Arrays.stream(Denomination.values())
-                .forEach(denomination -> cards.add(new Card(denomination, suit))));
+            .forEach(suit -> matchDenomination(cards, suit));
         Collections.shuffle(cards);
-        return cards;
+        return new ArrayDeque<>(cards);
+    }
+
+    private void matchDenomination(List<Card> cards, Suit suit) {
+        Arrays.stream(Denomination.values())
+            .forEach(denomination -> cards.add(new Card(denomination, suit)));
     }
 
     public List<Card> pickInitialCards() {
-        return Stream.<Card>builder()
-            .add(cards.pop()).add(cards.pop())
-            .build()
+        return Stream.generate(cards::pop)
+            .limit(2)
             .collect(Collectors.toList());
     }
