+++ src/test/java/blackjack/domain/user/HandTest.java
@@ -1,10 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Denomination;
 import blackjack.domain.card.Suit;
-import blackjack.domain.user.Hand;
-import blackjack.domain.user.HandStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
